uop: move apply_movement_op to movement, breaking schedule cycle
